+++ controllers/auth/index.js
@@ -1,13 +1,13 @@
-const signup = require("./signup");
-const signin = require("./signin");
-const getCurrentUser = require("./getCurrentUser");
-const logout = require("./logout");
-const verifyEmail = require("./verifyEmail")
+// const signup = require("./signup");
+// const signin = require("./signin");
+// const getCurrentUser = require("./getCurrentUser");
+// const logout = require("./logout");
+// const verifyEmail = require("./verifyEmail")
 
-module.exports = {
-  signup,
-  signin,
-  getCurrentUser,
-  logout,
-  verifyEmail,
-};
+// module.exports = {
+//   signup,
+//   signin,
+//   getCurrentUser,
+//   logout,
+//   verifyEmail,
+// };
