+++ controllers/transactions/index.js
@@ -1,0 +1,5 @@
+const addExpense = require("./addExpense");
+
+module.exports = {
+  addExpense,
+};
