+++ helpers/ctrlWrapper.js
@@ -1,11 +1,11 @@
-const ctrlWrapper = (ctrl) => {
-  return async (req, res, next) => {
-    try {
-      await ctrl(req, res, next);
-    } catch (error) {
-      next(error);
-    }
-  };
-};
+// const ctrlWrapper = (ctrl) => {
+//   return async (req, res, next) => {
+//     try {
+//       await ctrl(req, res, next);
+//     } catch (error) {
+//       next(error);
+//     }
+//   };
+// };
 
-module.exports = ctrlWrapper;
+// module.exports = ctrlWrapper;
