+++ helpers/index.js
@@ -1,4 +1,3 @@
+// const ctrlWrapper = require("./ctrlWrapper");
 
-const ctrlWrapper = require("./ctrlWrapper");
-
-module.exports = { ctrlWrapper };
+// module.exports = { ctrlWrapper };
