--- models/schemas/transaction.js
@@ -1,8 +1,0 @@
-// дата
-// описание
-// категория (нужен id)
-// сумма
-
-// к какой категории относится: расход или доход
-// какому юзеру принадлежит
-// индекс
