+++ models/transactions.js
@@ -1,0 +1,8 @@
+const { model } = require("mongoose");
+
+const { transactions } = require("./schemas");
+const { expenseSchema } = transactions;
+
+const Transaction = model("expense", expenseSchema);
+
+module.exports = Transaction;
