+++ models/user.js
@@ -1,8 +1,8 @@
-const { model } = require("mongoose");
+// const { model } = require("mongoose");
 
-const { user } = require("./schemas");
-const { userSchema } = user;
+// const { user } = require("./schemas");
+// const { userSchema } = user;
 
-const User = model("user", userSchema);
+// const User = model("user", userSchema);
 
-module.exports = User;
+// module.exports = User;
