+++ routes/api/transactions.js
@@ -1,2 +1,12 @@
 // Здесь должны быть роуты для получения отчета.
 
+const express = require("express");
+
+const { transactions: ctrl } = require("../../controllers");
+// const { authenticate } = require("../middlewares");
+
+const router = express.Router();
+
+router.post("/", ctrl.addExpense);
+
+module.exports = router;
