+++ routes/api/user.js
@@ -1,12 +1,9 @@
-const express = require("express");
+// const express = require("express");
 
+// const {
+//   user: { joiSchema },
+// } = require("../../models/schemas");
 
-const {
-  user: { joiSchema },
-} = require("../../models/schemas");
+// const router = express.Router();
 
-const router = express.Router();
-
-
-
-module.exports = router;
+// module.exports = router;
