+++ services/index.js
@@ -1,5 +1,7 @@
-const users = require("./users");
+// const users = require("./users");
+const transactions = require("./transactions");
 
 module.exports = {
-  users,
+  // users,
+  transactions,
 };
