+++ services/transactions.js
@@ -1,0 +1,9 @@
+const { Transaction } = require("../models");
+
+const addExpense = (newExpense) => {
+  return Transaction.create(newExpense);
+};
+
+module.exports = {
+  addExpense,
+};
