+++ utils/index.js
@@ -1,7 +1,7 @@
-const sendEmail = require("./sendEmail");
-const createEmail = require("./createEmail");
+// const sendEmail = require("./sendEmail");
+// const createEmail = require("./createEmail");
 
-module.exports = {
-  sendEmail,
-  createEmail,
-};
+// module.exports = {
+//   sendEmail,
+//   createEmail,
+// };
